refactor(handlers): extract query.ts handler from intent.ts (Stage B)
handleSAPQuery + SQL IN-list chunking/parsing. Relocates hasSqlParserSignature
to shared.ts (used by query + the dispatch error tree). 3672 tests, A1 zero-diff.
